refact: DashBoardViewModelをインターフェイスに変換した
